\encoding{UTF-8}

\name{print.Tinflex}
\alias{print.Tinflex}
\alias{print.TinflexC}

\title{
  Print Tinflex Generator Objects
}

\description{
  Print methods for generator objects of class \code{"Tinflex"} or
  \code{"TinflexC"}.
}

\usage{
\S3method{print}{Tinflex}(x, debug=FALSE, \dots)
\S3method{print}{TinflexC}(x, debug=FALSE, \dots)
}

\arguments{
  \item{x}{an object of class \code{"Tinflex"} or \code{"TinflexC"}.}
  \item{debug}{enable/disable the display of detailed information
    about the object. (logical)}
  \item{\dots}{additional arguments to \code{\link{print}}.}
}

\details{
  These are the \code{\link{print}} methods for objects of classes
  \code{"Tinflex"} and \code{"TinflexC"}.
}

\author{
  Josef Leydold \email{josef.leydold@wu.ac.at}, Carsten Botts and
  Wolfgang \enc{Hörmann}{Hoermann}.
}

\seealso{
  \code{\link{print}}. 
  \code{\link{Tinflex.setup}}.
  \code{\link{Tinflex.setup.C}}.
  See \code{\link{Tinflex.setup}} for examples.
}

%% \examples{}

\keyword{distribution}
\keyword{datagen}
