\encoding{UTF-8}

\name{print.Tinflex}
\alias{print.Tinflex}

\title{
  Print Tinflex Generator Object
}

\description{
  Print method for generator objects of class \code{"Tinflex"}.
}

\usage{
\S3method{print}{Tinflex}(x, debug=FALSE, \dots)
}

\arguments{
  \item{x}{an object of class \code{"Tinflex"}.}
  \item{debug}{enable/disable the display of detailed information
    about the object. (logical)}
  \item{\dots}{additional arguments to \code{\link{print}}.}
}

\details{
  This is the \code{\link{print}} method for objects of class
  \code{"Tinflex"}.
}

\author{
  Josef Leydold \email{josef.leydold@wu.ac.at}, Carsten Botts and
  Wolfgang \enc{Hörmann}{Hoermann}.
}

\seealso{
  \code{\link{print}}. 
  \code{\link{Tinflex.setup}}.
  See \code{\link{Tinflex.setup}} for examples.
}

%% \examples{}

\keyword{distribution}
\keyword{datagen}
