\encoding{UTF-8}

\name{Tinflex.sample}
\alias{Tinflex.sample}

\title{
  Draw Random Sample from Tinflex Generator Object 
}

\description{
  Draw a random sample from a generator object of class \code{"Tinflex"}.
}

\usage{
Tinflex.sample(gen, n=1)
}

\arguments{
  \item{gen}{an object of class \code{"Tinflex"}.}
  \item{n}{sample size. (integer)}
}

\author{
  Josef Leydold \email{josef.leydold@wu.ac.at}, Carsten Botts and
  Wolfgang \enc{Hörmann}{Hoermann}.
}

\seealso{
  See \code{\link{Tinflex.setup}} for examples.
}

%% \examples{}

\keyword{distribution}
\keyword{datagen}
